+++ src/backend/mod.rs
@@ -1,0 +1,3 @@
+//! Repository backends
+
+pub mod git2;